Fix problems with PPA experiment explanation (#9)
* Remove uneeded escaping

There appeared to be remnants of both HTML and Markdown escaping, which
resulted in invalid code in the code blocks.

* Fix misspelled property

`sources` is the spelling used in both the WebIDL description and in the
second JavaScript example, so `source` in the first JavaScript example
is likely incorrect.

* Remove unneeded quotes

As far as I am aware, `size` does not need to be quoted when used as a
property in an object literal. If it were, then it would need to be
quoted in the other example.

* Fix indentation

The start of this `dictionary` is difficult to see with the first line
indented.

* Make formatting more consistent

Both calls to `measureConversion` now have one property per line and use
trailing commas.

* Sort properties consistently

It is easier to read examples when the properties are written in the
same order everywhere.

* Fix indices used in example

The indices declared in the first table did not match up with the
indices used later in the example.